bridge/requirements.txt
requests
supabase
python-dotenv
python-telegram-bot
